+++ src/main/java/cart/application/ProductService.java
@@ -1,10 +1,10 @@
 package cart.application;
 
 import cart.dao.CartItemDao;
-import cart.domain.Product;
 import cart.dao.ProductDao;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.domain.Product;
+import cart.dto.request.ProductRequest;
+import cart.dto.response.ProductResponse;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
