+++ src/main/java/cart/application/auth/Auth.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.application.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
