+++ src/main/java/cart/dto/request/OrderCreateRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.request;
 
 import java.util.ArrayList;
 import java.util.List;
