+++ src/main/java/cart/dto/response/CartPointsResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 public class CartPointsResponse {
     private final int savingRate;
