+++ src/main/java/cart/ui/OrderController.java
@@ -1,10 +1,11 @@
 package cart.ui;
 
 import cart.application.OrderService;
+import cart.application.auth.Auth;
 import cart.domain.Member;
-import cart.dto.CartPointsResponse;
-import cart.dto.OrderCreateRequest;
-import cart.dto.OrderResponse;
+import cart.dto.request.OrderCreateRequest;
+import cart.dto.response.CartPointsResponse;
+import cart.dto.response.OrderResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
