+++ src/main/java/cart/ui/ProfileController.java
@@ -1,8 +1,8 @@
 package cart.ui;
 
-import cart.dao.MemberDao;
+import cart.application.auth.Auth;
 import cart.domain.Member;
-import cart.dto.ProfileResponse;
+import cart.dto.response.ProfileResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
